+++ src/BabylonCpp/include/babylon/collisions/collision_cache.h
@@ -18,19 +18,19 @@ public:
   CollisionCache();
   ~CollisionCache();
 
-  std::unordered_map<unsigned int, SerializedMesh>& getMeshes();
+  std::unordered_map<size_t, SerializedMesh>& getMeshes();
   std::unordered_map<std::string, SerializedGeometry>& getGeometries();
   bool containsMesh(unsigned int id) const;
   SerializedMesh& getMesh(unsigned int id);
   void addMesh(const SerializedMesh& mesh);
-  void removeMesh(unsigned int uniqueId);
+  void removeMesh(size_t uniqueId);
   bool containsGeometry(const std::string& id) const;
   SerializedGeometry& getGeometry(const std::string& id);
   void addGeometry(const SerializedGeometry& geometry);
   void removeGeometry(const std::string& id);
 
 private:
-  std::unordered_map<unsigned int, SerializedMesh> _meshes;
+  std::unordered_map<size_t, SerializedMesh> _meshes;
   std::unordered_map<std::string, SerializedGeometry> _geometries;
 
 }; // end of class CollisionCache
